+++ astropy/coordinates/builtin_frames/supergalactic_transforms.py
@@ -3,7 +3,8 @@
 from __future__ import (absolute_import, unicode_literals, division,
                         print_function)
 
-from ..angles import rotation_matrix
+from ..matrix_utilities import (rotation_matrix,
+                                matrix_product, matrix_transpose)
 from ..baseframe import frame_transform_graph
 from ..transformations import StaticMatrixTransform
 
@@ -16,9 +17,9 @@ def gal_to_supergal():
     mat1 = rotation_matrix(90, 'z')
     mat2 = rotation_matrix(90 - Supergalactic._nsgp_gal.b.degree, 'y')
     mat3 = rotation_matrix(Supergalactic._nsgp_gal.l.degree, 'z')
-    return mat1 * mat2 * mat3
+    return matrix_product(mat1, mat2, mat3)
 
 
 @frame_transform_graph.transform(StaticMatrixTransform, Supergalactic, Galactic)
 def supergal_to_gal():
-    return gal_to_supergal().T
+    return matrix_transpose(gal_to_supergal())