+++ astropy/utils/compat/numpy/__init__.py
@@ -8,3 +8,4 @@ in all supported NumPy versions.  See docs/utils/numpy.rst for details.
 from __future__ import division, absolute_import, unicode_literals
 
 from .lib.stride_tricks import broadcast_arrays, broadcast_to
+from .core.multiarray import matmul